code/python/pyproject.toml
[project]
name = "kontexto"
version = "0.1.0"
description = "Context-based word game in german."
authors = [
    {name = "paul",email = "paulkaisr@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]


[build-system]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [
    { include = "src" }
]
